+++ app/src/main/java/umc/everyones/lck/domain/model/request/mypage/CancelParticipateViewingPartyMypageModel.kt
@@ -4,7 +4,7 @@ import umc.everyones.lck.data.dto.request.mypage.CancelParticipateViewingPartyMy
 import java.io.Serializable
 
 data class CancelParticipateViewingPartyMypageModel(
-    val viewingPartyId: Int
+    val viewingPartyId: Long
 ): Serializable {
     fun toCancelParticipateViewingPartyMypageRequestDto() =
         CancelParticipateViewingPartyMypageRequestDto(viewingPartyId)
